[TRTLLM-14329][perf] Enable fused DSA metadata by default
